Fixes text encoding issues in output handling during clones
When a dataset name includes special characters there can sometimes
be issues handling kart clone output that causes error messages
to be displayed and the repos are not added to the manager.

Despite this, the repository is cloned successfully on disk and can
be manually added.

This change allows for unexpected encoding issues that don't impact
on the normal operation of the plugin to be ignored.